pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hellodata_pod_operator_params"
version = "0.0.14"
authors = [
  { name="Felix Neidhart", email="felix.neidhart@bedag.ch" },
  { name="Bedag Informatik AG", email="hellodata@bedag.ch" }
]
description = "A small library providing a way to pass parameters to a pod operator in Airflow."
readme = "README_PYPI.md"
requires-python = ">=3.8"
classifiers = [
    "Operating System :: OS Independent",
    'Development Status :: 1 - Planning',
    'Intended Audience :: Science/Research',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
]
license = {file="LICENSE"}
dependencies = [
    'apache-airflow-providers-cncf-kubernetes>=7.13.0',
    'kubernetes>=23.6.0',
]

[project.urls]
Homepage = "https://github.com/bedag/pod_operator_params"
Issues = "https://github.com/bedag/pod_operator_params/issues"